Transfer mock data from PlannerService to Database
